Separated component, added SVG to input
